refactor `cmd` package into package `cli`
